Updating from upstream source

+++ WDL/dirscan.h
@@ -68,14 +68,14 @@ class WDL_DirScan
 #endif
      ) // returns 0 if success
     {
-      int l=strlen(dirname);
+      size_t l=strlen(dirname);
       if (l < 1) return -1;
 
       WDL_String scanstr(dirname);
 #ifdef _WIN32
       if (!isExactSpec) 
 #endif
-	 if (dirname[l-1] == '\\' || dirname[l-1] == '/')  scanstr.SetLen(l-1);
+	 if (dirname[l-1] == '\\' || dirname[l-1] == '/')  scanstr.SetLen((int)l-1);
    
    m_leading_path.Set(scanstr.Get());
